iio: axi-adc: move DMA code into the core driver
It looks like the old code was in a different file, so that it's shared
with the other 4 AXI ADC clones.
Since those now have their own calls to DMA, and those are self-contained,
we can now move the 'cf_axi_adc_ring_stream.c' file into the core driver.
