+++ drivers/iio/adc/Makefile
@@ -29,7 +29,7 @@ obj-$(CONFIG_AD799X) += ad799x.o
 obj-$(CONFIG_AD9963) += ad9963.o
 obj-$(CONFIG_ADM1177) += adm1177.o
 
-cf_axi_adc-y := cf_axi_adc_core.o cf_axi_adc_ring_stream.o
+cf_axi_adc-y := cf_axi_adc_core.o
 obj-$(CONFIG_CF_AXI_ADC) += cf_axi_adc.o
 
 ad9208_drv-y := ad9208.o \
